chore(origin-ui-core): remove exports of dependent components and move the into the folders of its parent
BREAKING CHANGE: UploadChip and Dot components are no longer exported from origin-ui-core
