--- packages/ui/libs/ui/core/src/components/file/index.ts
@@ -1,1 +1,0 @@
-export * from './UploadChip';
