--- packages/ui/libs/ui/core/src/components/icons/Dot/Dot.stories.tsx
@@ -1,10 +1,0 @@
-import { Dot, DotProps } from './Dot';
-import { Meta } from '@storybook/react';
-import React from 'react';
-
-export default {
-  title: 'Utils / Dot',
-  component: Dot,
-} as Meta;
-
-export const Default = (args: DotProps) => <Dot {...args} />;
