+++ packages/ui/libs/ui/core/src/components/icons/index.ts
@@ -1,6 +1,5 @@
 export * from './IconLink';
-export * from './Dot';
 export * from './IconText';
-export * from './FallbackIcon';
+export * from './IconHoverText';
 export * from './IconPopover';
 export * from './FilterIcon';
