+++ packages/ui/libs/ui/core/src/components/text/TextWithPendingDot/Dot/Dot.tsx
