+++ packages/ui/libs/ui/core/src/components/text/TextWithPendingDot/Dot/index.ts
