+++ packages/ui/libs/ui/core/src/components/text/TextWithPendingDot/TextWithPendingDot.tsx
@@ -1,12 +1,12 @@
 import React, { FC } from 'react';
 import { Tooltip, Typography, TypographyProps } from '@mui/material';
-import { Dot } from '../../icons';
+import { Dot } from './Dot';
 import { useStyles } from './TextWithPendingDot.styles';
 import { useTheme } from '@mui/styles';
 
 export interface TextWithPendingDotProps {
   textContent: string;
-  pending: boolean;
+  pending?: boolean;
   tooltipText?: string;
   typographyProps?: TypographyProps & { component?: 'span' | 'p' };
   showSuccessDot?: boolean;
@@ -15,12 +15,14 @@ export interface TextWithPendingDotProps {
 export const TextWithPendingDot: FC<TextWithPendingDotProps> = ({
   textContent,
   pending = false,
-  tooltipText,
+  tooltipText = '',
   typographyProps,
   showSuccessDot = false,
 }) => {
   const classes = useStyles();
   const theme = useTheme();
+  const dotBgColor = showSuccessDot ? theme.palette.primary.main : undefined;
+
   return (
     <div className={classes.blockWrapper}>
       <div>
@@ -29,11 +31,7 @@ export const TextWithPendingDot: FC<TextWithPendingDotProps> = ({
       {(pending || showSuccessDot) && (
         <Tooltip title={tooltipText}>
           <div className={classes.dotWrapper}>
-            <Dot
-              backgroundColor={
-                showSuccessDot ? theme.palette.primary.main : undefined
-              }
-            />
+            <Dot backgroundColor={dotBgColor} />
           </div>
         </Tooltip>
       )}
